shut compiler warning by casting a mode_t to int when used in printf. mode_t is an integer.

+++ cmds.c
@@ -588,7 +588,7 @@ cmd_lumask(int argc, char **argv)
 		/* show current umask */
 		modetmp = umask(modeold), umask(modetmp);
 		modetmp &= S_IRWXU|S_IRWXG|S_IRWXO;
-		printf("%03o\n", modetmp);
+		printf("%03o\n", (int)modetmp);
 		break;
 	case 2:
 		/* set new umask */
